logcli/pyproject.toml
[project]
name = "logcli"
version = "0.1.0"
description = "Log Analyzer & Monitoring CLI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3",
]
